% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{SticsRPacks_deps}
\alias{SticsRPacks_deps}
\title{List all SticsRPacks dependencies}
\usage{
SticsRPacks_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
SticsRPacks packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
List all SticsRPacks dependencies
}
