% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{SticsRPacks_logo}
\alias{SticsRPacks_logo}
\title{The SticsRPacks logo, using ASCII or Unicode characters}
\usage{
SticsRPacks_logo(unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\description{
Use \code{\link[crayon:strip_style]{crayon::strip_style()}} to get rid of the colors.
}
\examples{
SticsRPacks_logo()
}
