% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SticsRPacks_packages}
\alias{SticsRPacks_packages}
\title{List all packages in SticsRPacks}
\usage{
SticsRPacks_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Logical, TRUE to include the SticsRPacks package
FALSE otherwise}
}
\description{
List all packages in SticsRPacks
}
\examples{
SticsRPacks_packages()
}
