% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{SticsRPacks_update}
\alias{SticsRPacks_update}
\title{Update SticsRPacks packages}
\usage{
SticsRPacks_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
SticsRPacks packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
This will check to see if all SticsRPacks packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
SticsRPacks_update()
}
}
