% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SticsRPacks.R
\docType{package}
\name{SticsRPacks-package}
\alias{SticsRPacks}
\alias{SticsRPacks-package}
\title{SticsRPacks: Easily Install and Load 'SticsRPacks' packages}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

'SticsRPacks' is a set of packages that work in
    harmony because they share common data representations and 'API'
    design. This package is designed to make it easy to install and load
    multiple 'SticsRPacks' packages in a single step. Learn more about
    'SticsRPacks' at <https://sticsrpacks.github.io/SticsRPacks/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sticsrpacks.github.io/SticsRPacks}
  \item \url{https://github.com/SticsRPacks/SticsRPacks}
  \item Report bugs at \url{https://github.com/SticsRPacks/SticsRPacks/issues}
}

}
\author{
\strong{Maintainer}: Rémi Vezy \email{remi.vezy@cirad.fr}

Authors:
\itemize{
  \item Samuel Buis \email{samuel.buis@inra.fr}
  \item Patrice Le Charpentier \email{patrice.lecharpentier@inra.fr}
  \item Michel Giner \email{michel.giner@cirad.fr}
}

Other contributors:
\itemize{
  \item SticsRPacks [copyright holder, funder]
}

}
\keyword{internal}
