% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{SticsRPacks_update}
\alias{SticsRPacks_update}
\title{Update SticsRPacks packages}
\usage{
SticsRPacks_update(
  ref = list(SticsRFiles = NULL, CroptimizR = NULL, SticsOnR = NULL)
)
}
\arguments{
\item{ref}{The desired git reference. Could be a commit, tag, or branch name. Defaults to NULL
for the last release of the package.}
}
\description{
This will check to see if all SticsRPacks packages are up-to-date, and install the desired version if not.
}
\examples{
\dontrun{
# Use the last release:
SticsRPacks_update()

# Using a particular commit for one:
SticsRPacks_update(ref = list(SticsRFiles= "42f9333e1b1c336cf431e2e33dc13caa994a3ae9")
# Or a version:
SticsRPacks_update(ref = list(SticsRFiles= "v0.1.0.9003")

}
}
